Vectors Implements
